+++ src/main/java/com/pyding/deathlyhallows/blocks/BlockVisConverter.java
@@ -1,6 +1,6 @@
 package com.pyding.deathlyhallows.blocks;
 
-import com.pyding.deathlyhallows.blocks.tiles.VisConverterTile;
+import com.pyding.deathlyhallows.blocks.tiles.TileEntityVisConverter;
 import net.minecraft.block.BlockContainer;
 import net.minecraft.block.material.Material;
 import net.minecraft.tileentity.TileEntity;
@@ -13,7 +13,7 @@ public class BlockVisConverter extends BlockContainer {
 
 	@Override
 	public TileEntity createNewTileEntity(World world, int metadata) {
-		return new VisConverterTile();
+		return new TileEntityVisConverter();
 	}
 
 	@Override
